Crate alglobo

AlGlobo.com - Process payments

This program sets up N_NODES which will process all of the payments from the payments csv file. It will send the request to all the designated agents, using the configuration in the agents.yaml file.

If the leader node is killed, another one is elected using the ring election agorithm.

Start the program with cargo run --bin alglobo <payments_file>.csv (or default to a csv if not provided)

Modules

alglobo_node

AlgloboNode struct

communication

Several definitions regarding the alglobo<->agents communication protocol

logger

Logging functions

utils

Global utils related to reading and parsing of csv and yaml files

Functions

main

Starts the main process, starting the node killer and each node process

psycho_node_killer

Starts the thread designated to kill each node via keyboard input.