Module alglobo::airlines

Handle airlines config

Structs

Airline

WebServer actor airline

Constants

DEFAULT_RETRY_SECONDS

Functions

from_file

Create a Airline Server for each available airline in filename. Each Server allows to launch multiple instances of the specific Actor Airline according to the rate limit of the server. Returns a HashMap with the name of the airline and the address of the server.

Type Definitions

Airlines