Module alglobo::alglobo_node

AlgloboNode struct

Handles the logic of each Algobo.com node, where the leader is chosen with the ring election algorithm. The leader will send each transaction result to the replicas.

The communication between nodes is done via UDP, where each node will have two sockets. One for receiving leader election messages and the other for receiving transaction results.

If the replicas reach a timeout without messages from the leader, a new leader is coordinated.

Structs

AlgloboNode

AlgloboNode struct

Constants

MSG_ACK

Control message for ACKs

MSG_COORDINATOR

Control message for coordinator messages

MSG_ELECTION

Control message for election messages

MSG_KILL

Control message for kill messages

N_NODES

The amount of nodes to launch the program with

TIMEOUT

Timeout for receiving transaction information in the replicas

Functions

id_to_ctrladdr

Socket used for receiving leader election/coordination messages

id_to_dataaddr

Socket used for receiving transaction information from the leader